Abstract out penalty zero/assemble methods
* I am going to start pouring in the penatnlty and
flux capculations and wanted to avoid multiple zero
and assembly calls. This was due to the fact that the
algorithms had a generic driver..

* Add AssembleNonConformalAlgorithmDriver class to manage
pre and post in a general manner. Made the change and all
tests pass. However, I noticed that the parallel assembly for
normal flux was missing..

* Add parallel assembly for normal flux in newly generalized
AlgDriver. These were missing in both edge and element.

This causes diffs in the dg-based rtests.