Change Debug to case sensitive DEBUG
* also, I am managing debug and opt builds better now - at least between
Trilinos and Nalu. Basically, I created a new build_debug in both Trlinos
and Nalu. Then, I customized the do-config file to pull in the correct
opt or debug location. The install for trilinos points to trilinos_debug and
the Nalu build points to this trilinos install location. Now, I can simply
switch between the build and build_dbg.

Most of these files are "untracked".