Tet shifted bug; deal with Pyr for shifted
* shifted shape functions ip locations were off for tet.

* deal with pyr. The pyr shape functions are not linear along
an edge. Therefore, we will not expect to see this shifted
appraoch provide the same result from a, e.g., pry and tet side
sharing the same edge. There are fixes for that which I will evaluate.

In the meantime, the shifted hyrbid CVFEM mesh sim now works.. added a test
so that when I change it (fix it) I know for sure good coverage exists.