Deploy NC (and CHT) for hex/tet topos
* CHT requires isInElem for 3DQuad, 3DTri, and 2DEdge. This is done
now with the addition of the 3DTri methods.

* NC alg requires inInElem for the above.

* NC alg requires general_face_grad_op which is now coded for
HexSCS, TetSCS, Tri2DSCS and Quad2DSCS. Wedge and pyramids remain,
however, will probably not be coded soon.

* NC alg requires mapping for side coords to elem coords. This is
complete for HexSCS, TetSCS, Tri2DSCS and Quad2DSCS. Same deal with
pyramids and wedges..

InterpolatePoint is not implemented for every topo... Tet, Pyr and Wedge
missing...