Remove storing of sc_volume
* In low order Nalu, we never use the sc_volume (gauss point sub-control
volume). In higher order, storing this data is problematic. Therefore,
remove storing in favior of a on-the-fly re-calculation.

Notes:

a) for some systems, we load up the ElemSuppAlgs. Therefore, we have
redundant gathers and calculations of dn/dx and ws_scv_volume. The helium
plume is such a case. Therefore, timings slightly increased. I think that
this can be fixed by adding higher level notions of gather and geometry
calculations... For now, the cost is ~4%.