Clean up reduced options at continuity LHS algs
* When the user requests a shifted poisson operator, then reduced sens are enforced.
In this case, dndx should only be required to be computed once. This code path was
not actually occuring before. Therefore, shifting the CVFEM operator changed the RHS
however, the LHS was in a strage state in that the full sensitivities were being used.

* When the user requests only reduced sensitivies, then two ws scratch
spaces are required: dndx (non-shifted Poisson) and dndx_lhs (reduced/shifted).

* Added proper LHS appraoch at open bcs. The former code was not consistent at open
bcs.