Remove old design for turbulence averaging in favor of new volume/surface approach
New syntax:

    turbulence_averaging:
      time_filter_interval: 10.0
      forced_reset: no

      specifications:
        - name: one
          target_name: block_1
          reynolds_averaged_variables:
            - velocity
            - mixture_fraction
          favre_averaged_variables:
            - velocity
            - mixture_fraction
          compute_reynolds_stress: no # not yet active
          compute_tke: no # not yet active

        - name: two
          target_name: surface_5
          reynolds_averaged_variables:
            - normal_heat_flux
          favre_averaged_variables:
            - normal_heat_flux

Output of fields is as follows: primitive_name + _ra/fa_ + name, e.g.,

velocity_ra_one, velocity_fa_one, normal_heat_flux_ra_two, normal_heat_flux_fa_two.

Averaging review provides this overview of fields registered and to be processed. Note that
if two blocks are included, then one would end up with two reynolds averaged density fields.

* major refac; now use look-aheads to create post processing classes. I find this a preferred
approach as compared to weaving in field registration off of the equation systems.

* remove old averaging infrastructure; may be no need for AveragingInfo since I can store
off the YAML node?

* deal with possible missing lines above, e.g., reynolds_averaged_variables and farve_averaged_variables
by extracting a pointer and not assuming that it is there (YAML throw)

* velocity_fa/ra is a special field registration case while everything else uses a
generic field registration interface. This retains the vector notion of velocity.

* clarify an internal naming convention

Notes:

a) this is totally a new style of dealing with these types of post-processing events.
First, the objects are created only when the parsed line is found. Also, the parsing
needs to specify the set of target parts. This allows one to free nodal registration
to the post processing class itself. SolutionNorm and WallFunction PP are candidates for
the same approach...This requires the user to know the target parts.

b) averages are still only nodal and double in type... Later, I can generalize this.

c) one can not mix and match field types. For example, adding normal_heat_flux to "one" above
(a volume block) would result in an error that the field can not be found at setup time.

d) painful diffing to make sure that the new methodology is working in restart, multiple blocks,
etc.