Support input/output use cases for data transfer
* Create a new Realm, InputOutputRealm, which is derived
from Realm. This realm simply loads the mesh and registers
fields. Transfers are expected as follows:

transfers:

  - name: xfer_fluid_io
    type: geometric
    realm_pair: [fluidRealm, ioRealm]
    mesh_part_pair: [block_1, block_1]
    objective: input_output
    transfer_variables:
      - [mixture_fraction, mixture_fraction]

  - name: ioRealm
    mesh:  mesh/io_mesh.g
    type: io

    field_registration:
      specifications:
        - field_name: mixture_fraction
          target_name: block_1
          field_size: 1
          field_type: node_rank

        - field_name: reynolds_stress
          target_name: block_1
          field_size: 6
          field_type: node_rank

    output:
      output_data_base_name: output/io_results.e
      output_frequency: 2
      output_node_set: no
      output_variables:
       - mixture_fraction
       - reynolds_stress

* partition types of transfers, e.g., multi_physics, input_output
and initialization.

* provide norm and incremement only when multiphysics

* helium plume simulation works:)

* changed adaptive time step return to 1e8

* very odd tiniest of tiny diff in the adaptive time step - only on
step 3 of the simulaiton... Everything else matches perfectly. valgrind
shows zero memory leaks. No obviuous code issue, however, when
xfers are turned off, the test fully passes. Nonlinear residuals are
clean; it is only the adaptive tme step that differs in the last diget.

Notes:

a) Okay, I do not think that this is so bad of a design..
b) plumbing for IC should all be there and should work... will test next