Use case for initial condition from mesh file
* Rework some small stuff here and there

* erase Realm after processing the initial condition

* annoying... TKE was interpolated to be negative and, therefore,
generating nans since the field was very near zero; add a clip
at the initialize phase.

* add pre_work for tke and SST that clips if an IC is negative..

* rename some transfer calls to account for the three types:
multi_physics, initialization and input_output. This makes the
code more clear.

* remove delete and erase of initialziation region type.. WIP, I guess
until I have shared pointers.. The issue is that the Realms and TimeIntegrator
both have a vector of Realm. I could have checked for NULL in the Realms
destructor, however, that seemed like a cludge in that the vector on
Realms was out-of-date.

Notes:

a) noticed that the gold file for elemBackLRSST is 1e-11 off from current.
However, current matches previous nightly.

b) helium norm is back as it was yesterday!