Add flow past cylinder temperature bc
* add infrastructure for temperature to be provided at inflow, open
and wall bcs.

* add new FlowPastCylindertempAuxAlg to search for temeperature given
wall position and, in turn, the angle, theta.

* modified temperature_bc_setup to allow for functions. Since we pass in
UserData, I needed to add temperature to the UserData base class... Consider
templating on UserData type to fix this...

* add FlowPast files

Notes:

a) some issues with interpolation routine... Index in search is not
properly found; fixed by iMid < iMin... Works in my stand-alone code and
in spot checking the rtest