Clean up xfer
* removed derivation of Transfers on std::vector in favor of the
previously deployed design of holding a std::vector of objects, e.g.,
Realms, MaterialPropertys, etc.

* changed Z and h at open for PNG; use dof value, not user specified value.
This is actually a bit more complex at open bcs in that the specified
value is used only at entrainment. Safest thing is to use local dof value as
this removes all of the "crud" at the open bcs when using DCO + 4th order.

* user dials in output_clipping_diagnostic for Mixture fraction and
enthalpy... Defaults are false for Z; true for h