Bug fix in Tpetra with multiple periodic (thanks to SK)
Periodic boundary conditions in Nalu are implemented by a "pseudo-node"
approach: each slave node knows the global identifier of its' master, and
only master (and regular) nodes contribute rows to the system matrix.
When assembling the right-hand-side or matrix, contributions to slave nodes
are redirected to the rows/columns associated with their master nodes.
When there is more than one slave node referencing a master node, the code
was introducing a new local ID for each slave node.  This change checks if
the master node has already been visited and re-uses the existing local ID
(previously, multiple global ID/local ID pairs were sent to Tpetra which
could cause a failure).  The new code is a bit simpler in that it directly
sorts the global ID vectors instead of the vector of nodes, from which the
global ID vectors were then extracted.

* add TGV IC for more detailed testing of periodic bcs