Allow the user to specify search tolerance and search expansion factor.
The core issue here was that the default search tolerance (which is used
as a point radius was dimensional and set to 1e-4. Meshes fully resolved,
therefore, may have found many many boxes which increases ghosting overhead,
fine search, etc.

While I was here, I removed the hard coded expansion factor. This parameter
dials in the increase in radius when points are not found.

* new line commands:

  - name: xfer_fluid_io_volume
    search_tolerance: 5.0e-4
    search_expansion_factor: 1.45

Defaults are 1.5 and 1.0e-4