Add actuator line wip; deal with single sphere first per specification
Lots of placeholders for a real drag calculation. At present, the point drag
term is based on our Lagrangian particle model (spheres). The gaussian profile
is provided by the input file based on at what radisu do you want the factor to be.

The best element (or element that holds the precise point centroid) computes the drag
and then each element in the collection of elements in the bounding sphere is iterated
and assembles the nodal source term. LHS contributions are assembled only for the element
that holds the centroid of the point.

Each specification resides on a unique processor. The rank that owns the point is the
rank to which elements in the radius sphere is ghosted. A viable approach is the opposite,
 however, this works well given STK infrastructure.

* add ActuatorLine class; Realm manages this.

* add field registration in LowMachEQS

* hack in a drag source term model

* add source terms and LHS contribution

* add momentum assembly class

* put the field on the part!

* sign issue

Notes:

a) looks like it is working in serial and parallel.
b) feels like this may be a bit expensive... timing will tell...