Re-work search tolerance for periodic search
In realms with multiple periodic bc definitions, the intent of the
code is to use the most stringent search tolerance. However, this
tolerance was checking against the default tolerance - which
was really small - which would prevail in the end. Change default to
something large. Also, add iteration should points not be found.

* Modify periodic search attempts by changing search tolerance
on the fly.

* Hard coded Max attempts = 10; amplification factor = 5.

* Need to be careful with simply calling build_constraints() since
it assumes vectors are cleared. Add a finalize_search() method
that can be called many times within build_constraints().

No diffs now...

Notes:

Periodic Review:  realm: thermalRealm
=========================
Translating [ 5.83801e-20 -2.0016e-19 -0.1 ] Master/Slave pair
_____________________________________
Probable issue with Search on attempt: 1
the total number of slave nodes (832)
does not equal the product of the search (692224)
The algorithm will reduce the search tolerance: 0.02
_____________________________________
Probable issue with Search on attempt: 5
the total number of slave nodes (832)
does not equal the product of the search (2368)
The algorithm will reduce the search tolerance: 3.2e-05
---------------------------------------------------
Parallel consistency noted in master/slave pairings: 832/832
---------------------------------------------------