A straw-man proposal for initial unit-testing.
unit_tests.C is the main, and lives in the top-level nalu directory. The
build creates an executable at the same level as naluX/naluXd, which can
be run like this: % mpirun -np 1 unit_tests

Unit-test code lives in the unit_tests subdirectory. 1 unit-test is
in there now, SuperElement.C. That file contains a gtest 'TEST(...' at
the bottom of the file, which launches the SuperElement test code.

A future addition will be a convenience script which runs the unit_tests
executable for a few different processor-counts and reports pass/fail
overall.