Optimize the handling of ghost elements in NonConformalManager.
Added a function to NonConformalManager called
computePreciseGhostingLists. This function starts from the elemsToGhost_
vector, figures out which of those are already ghosted, and figures out
which receive-ghosts no longer need to be ghosted. Then the stk
change_ghosting function can be called with a minimal list of elements
to ghost, and a list of ghosts that can be removed from ghosting.

This change removes the need to destroy and re-ghost everything, and
also reduces the number of calls to stk modification_end.
For the test dgNonConformalEdgeCylinder this reduces runtime from
~71s to ~53s on linux blade.

I will also propose a patch to the stk team which further reduces
runtime for this test to ~49s. (The stk patch improves efficiency
within the stk ghosting operation.)