Add opposing unit normal calculation at precise GP
For the DG scheme, we require the opposing surface normal. Up
until this point, this was approximated by the first component
of the opposing exposed_area_vector. This was a hack, although
suitable for P=1 elements.

In this project, relax this assumption by adding a new
general_normal() so that when running higher order, we can use
the proper normal at the isoparametric coordinates for the
opposing face element.

* Add general_normal() for 3DQ4 quad faces

* Add general_normal() for 3DQ9 quad faces

* Add general_normal() for Edge2D bar faces; omit Edge92D

* Add general_normal() for Tri3D bar faces

* Modify all NonConf algs to gather ws_o_coordinates
and compute general_normal; not yet fully hooked up
to algorithm.

* hack in parallel max for difference between two normal calcs;
remove this as it was a temp test to see if the normals are looking
fine. For the most part, yes. However, it does seems that picking the
0th ip is different from using the local normal..

* complete the removal of opposing exposed area_vec

Notes:

a) minimal diffs. Now we are ready for HO ME method deployment.