Simplify manner in which nearest node is extracted for NonConfAlg
* Prior code for the NodalGradNonConfAlg used the meSCS->ipNodeMap[ordinal]
interface. This required one to load the element-node connectivity.

Alternatively, we should be able to use: meFC->ipNodeMap[0] and use
the already face-nodal connectivity.

This project makes that happen.

* when changing teh routine, it was noted that a diff was found. Therefore,
this pointed to an inconsistency between the above to approaches to obtain a
nearest node to an explosed ip. Testing showed that ordinal 2 was wrong
on the Hex27SCS ME. RK looked into it (is writing a unit test now) and found
an issue in an internal mapping.

Notes:

a) interestingly, only the new P1/P2 higher order test has an ordinal
of 2. The hoHelium test does not...