Final clarity in NSO code implementation.
* add back density scaling. Interesting story here... The theory manual
had the resiudal for KE to be: ~ u*Gjp/2. This was missing a density
scaling.

However, the expression for nu, in the KE form, should be:

	rho*sqrt(R*R/(dKE/dx*gij*dKE/dx)

not:

	sqrt(R*R/(dKE/dx*gij*dKE/dx).

In the end, the density scalings cancel. As such, the original code was
correct, however, it is far more clear to have the code represented as
above so that the KE resiudal is not confused as KE*rho residual. Density
should never be zero.

Notes:

a) nu in the NSO should always have units of kg/m^3/s.

b) The product nu*gij provides what appears to be a viscosity; kg/m/s.