Demonstrate unit-test fixture.
Made a class 'MasterElementHexSerial', which declares/creates the mesh,
then changed 'TEST' functions to 'TEST_F', which are actually member
methods of the MasterElementHexSerial class so they can refer to the
mesh and other members of that class.

In this case it saves a few lines of duplicated code for each test,
not huge, but demonstrates a handy gtest feature which can remove
a lot of duplication in some cases.