Fix restart output issue when mesh motion is active
* At a restart, we possibly output the first step of results. However,
mesh motion is applied at the integrate_realm() location. Although the
nonlinear residuals are fine with this setup, the output of the previous
results is screwed up since all of the mesh fields are zero. Simply
add a calculaiton of mesh motion quantities in the initial_work() method..

* added log output for when restart or output data files are written.

Notes:

a) add a restart test for mesh motion (threeBlades)
b) restart is not 100% correct; need to resolve this