Prototype algorithm design to remove resize
The objective of this project is to remove the usage of resize in the
algorithm::execute() in favor of a homogeneous collection of parts on which
the algorithm operates.

* Modify Algorithm class to manage a map<string, Algorithm*>

* LowMachEOS now creates a set of CourantReyAlgorithms for each
topology,

  // non-solver CFL alg
  const std::string cflAlgName = "cflAlgName_" + baseSCSName;
  std::map<std::string, Algorithm *>::iterator its
    = cflReyAlgDriver_->algorithmMap_.find(cflAlgName);
  if ( its == cflReyAlgDriver_->algorithmMap_.end() ) {
    AssembleCourantReynoldsElemAlgorithm *theAlg
      = new AssembleCourantReynoldsElemAlgorithm(realm_, part, meSCS);
    cflReyAlgDriver_->algorithmMap_[cflAlgName] = theAlg;
  }
  else {
    its->second->partVec_.push_back(part);
  }

The above results in multiple algorithms (each with many potential parts)
of the same topology. The core design would be that one algorithm type
holds all of the parts which can be heterogeneous.

* Rather than setting size in constructor and allocating space for member
data, save off master element, npe, nip and size in the execute. This is a
compromise between holding data and setting data in the execute.

* comment in a builder method concept

Notes:

a) kind of extensive - I think...
b) does this move towards the unit test approach to remove a realm?
c) Tricky... Algorithms, since they once handled all of the parts
   now operate on a possible subset of parts. As such, calls such as
   the following are tricky:

  // parallel max
  double g_maxCR[2]  = {};
  stk::ParallelMachine comm = NaluEnv::self().parallel_comm();
  stk::all_reduce_max(comm, maxCR, g_maxCR, 2);

  // sent to realm
  realm_.maxCourant_ = g_maxCR[0];
  realm_.maxReynolds_ = g_maxCR[1];

  You see, we now need something like the following:

  // before the execute
  realm_.maxCourant_ = 0.0;
  realm_.maxReynolds_ = 0.0;
  momentumEqSys_->cflReyAlgDriver_->execute();

  // sent to realm; at the end of the Alg class
  realm_.maxCourant_ = std::max(realm_.maxCourant_, g_maxCR[0]);
  realm_.maxReynolds_ = std::max(realm_.maxReynolds_, g_maxCR[1]);

  So... Conclusion is that Algorithms can no longer be extected to
  run without an AlgorithmManager to organize the pre and post-work.

d) what about edge-based algs. They really do need to be all unioned.
This requires a different interface to edge-based algorithm construction.