Add a unit-test of simulated element-algorithm using Supplemental Alg.
UnitTestElemSuppAlg.C contains a test (start reading at the bottom of
the file) which demonstrates an element-algorithm using a supplemental
alg, and that supplemental alg uses data from a 'ComputedElemData'
struct which allows for multiple supplemental algs to share data such
as gathered coordinates, scs_area_vector, etc., without each alg
redundantly computing that data. Each alg declares which data it needs,
so that the ComputeElemData struct only computes needed data.

The ComputedElemData struct is not templated. It contains arrays that
are sized using 'max' numbers, but it is only 1 element's worth of
data. However, a templated version of this struct could use precisely-
sized arrays.

The ComputedElemData struct is intended to be declared at the inner
loop (i.e., for each element entity) such that each thread can have
its own version of the arrays. This is necessary because other threads
will be operating on other elements so the data needs to be unique.
Also, if ComputedElemData contained read-write arrays, then they would
also need to be unique per thread to prevent write-conflicts.

Overall this is still a little crude, but we can iterate and refine
as appropriate.