Move forward, in earnest, with homogeneous supplemental algorithm design
* Modify ElemSolverAlgorithm to be homogeneous in nature.

* add new solverAlgorthmMap_ to SolverDriver. Specifically, a
<string, algorithm *> construct that allows for creating
algorithms not based on ALGORITHM_TYPE, rather,
std:: string AlgorithmName + "_TopoName". This allows for a
homogeneous collection of algorithms based on topology.

* Add CVFEM scalar diffusion; convert FEM use consolidated approach
to be homogeneous as well (rename class)

* Added a 3D MMS file to ensure that the old and new match **perfectly**

* add new cvfemHC test using consolidated design

Notes:

a) the MyClassFileDep.C is for "Dep"reacated...
b) at present, it is a delicate situation in that SuppAlgs need to
call either the non-master-element signature or the old style. We may
need to deprecate a few files for a while until this is dialed in
correctly.