Allow new use case for nonconformal part definition
* In some cases, the user may have defined multiple surface interactions.
Therefore, the current specification, e.g.,

    - non_conformal_boundary_condition: bc_in_out
      target_name: [surface_5, surface_7]

may not be the most convenient.

An alternative surface definition can now be as follows:

    - non_conformal_boundary_condition: bc_in_out
      current_target_name: surface_7
      opposing_target_name: [surface_5, surface_6]

In some cases, the user may also want to add multiple surfaces together, e.g.,
consider the three-blade example:

    - non_conformal_boundary_condition: bc_in_out
      current_target_name: [surface_8, surface_9, surface_10]
      opposing_target_name: [surface_88, surface_99, surface_1000]
      non_conformal_user_data:
        expand_box_percentage: 5.0
        search_tolerance: 0.01

Notes:

a) The rule remains the same in that each non_conformal_boundary_condition
block must contain a set of parts that fully cover the desired set
of surfaces. Otherwise, the search may fail with not-found pairs
to each integration point on the current surface.

b) a symmetric NC block is still expected. In theory, we could allow
for one bc definition and swap current and opposing behind the scnenes. However,
the code path looked a bit complex to do that now... A possible improvement
in the future, perhaps..