Add lumped mass for element-mass contribution
* allow for: lumed_*_time_derivative for ElemSuppAlg. In time,
we will be able to activate all SuppAlgs for the element-based scheme
to avoid a nodal assembly loop.

Notes:

a) tested for a hyrbid mesh (cylinder case) which I will push soon.

b) at present, a AssembleNodalAlgorithm is always required even if there is
no source term required. This is applicable to the element-based scheme.
A bug will be submitted soon to resolve this.

c) the code is now ready to try out the templated approach for consolidated
algorithms with all source terms as element-based.