Minor cleanups and optimizations in TpetraLinearSystem. (#79)
1. Moved globally-owned graph construction from
beginLinearSystemConstruction to finalizedLinearSystem (in preparation
for future refactoring of the graph construction process to use
exact profile information for a static graph).

2. Made lookup_myLID method take a const char* instead of a
const std::string& for a debugging string. Passing a string-literal
to a const std::string& causes a std::string construction, which was
sticking up on a profile.

3. Made addConnections take a pointer and length instead of a
std::vector. In many cases there is no need to construct and fill
a std::vector to pass in, when the data being passed is directly
available in a pointer. (i.e., no need to copy). Some cases still
require a copy, but those work just as well passing vector.data() and
size.

4. Clean up the looping in addConnections to match the upper-triangle-
connectivity that is being stored. (It was doing full looping. We
may go back to full looping at some point, depending on how the
graph-profile refactoring goes, but currently it is not needed.)

These optimizations are mostly minor, a couple percent of runtime
on the uqSlidingMeshDG case.