Add MomentumNSOElemSuppAlg that will work with new AssembleElemSolverAlg design
* refactor MomentumNSOElemSuppAlg.C. This is still a iffy design
since it requires a static source name per class and a hard coded set of
options. RK is looking into this. At present, hard coded to be the 4th order full
residual form. This may actually not be correct given that the flow has a buoyant
source term...

* change former SuppAlg to follow a standard that uses v_myField for local Kokkos
views and v_myPrivateField_ when views are data member.

Notes:

a) new class is not yet tested, so it may be buggy. SA and I are converging to a set of
capability.