Add error check for coincident nodes
* intersect surface parts to determine if coincident nodes are in existance.

* minor cleanup in NCAlgManager NCAlgInfo call

Notes:

a) Usage:

        - non_conformal:
            coincident_node_error_check: yes

b) If this error becomes difficult for the users to resolve, we can consider repairing the interface
by creating a new node with new unique id, changing all of the element/node connectivity and
managing parallel consistency. The algorithm would be very similar to the nodal promotion for
higher order.