Improve TpetraLinearSystem initialization and assembly.
Initialization is improved by computing an array of row-lengths
to give to CrsGraph at construction (instead of the previous hard-
coded '8').

Assembly is improved by calling sumIntoValues on the local kokkos
matrix obtained from the Tpetra::CrsMatrix object via getLocalMatrix().
Previously, we were calling Tpetra::CrsMatrix::sumIntoLocalValues.

Some other optimizations are done, such as reducing the number of calls to
CrsGraph::insertGlobalIndices by trying to process more of the
column-indices for a given row, rather than making separate calls
for each row-column pair.

Overall, these changes appear to speed up the uqSlidingMeshDG test
by ~10%. The cube6M problem running on serrano is sped up by about
an 8-10% improvement in the 'init' timer. The performance of the
Kokkos::CrsMatrix::sumIntoValues method appears to be the same as
that of the Tpetra::CrsMatrix::sumIntoLocalValues method, to within
noise. Some other tests, such as the nightly hoHelium test, appear
to be relatively un-affected by these changes.