Refactor MasterElement and SupplementalAlgorithm logic to improve modularity (#118)
* Refactor SuppAlg vector from Algorithm to SolverAlgorithm

Restrict the visibility of the vector holding SupplementalAlgorithm
instances to only classes that require it.

* Extract MasterElement creation logic from Realm into a standlone
function.

Creates two top-level standlone functions

    - get_surface_master_element
    - get_volume_master_element

that are intended to replace the equivalent methods in Realm. The
requested MasterElements are created only once and references to
previously created instances are returned on subsequent invocations.
Refactor all instances of MasterElement creation.  Remove the dependency
on Realm to create MasterElement instances.

* Revert "Refactor SuppAlg vector from Algorithm to SolverAlgorithm"

This reverts commit 26aa67f07e7685bc17e5f86b93a91db5cec7a425.