Transition wall function to a pure edge- or element-based implementation
* convert standard wall function flavor

* convert ABL wall function flavor

Notes:

a) The motivation of this push is to increment towards allowing the
element-based scheme to assemble the pressure gradient term within
interior and boundary contributions rather than applying a nodal source
term in time. With this commit, all adv/diff edge- and element-based matrix
contributions are in seperate implementations. The next step will be to add
a total advection term to Elem-based algoirthms, i.e., rho*ui*uj + p*del_ij