Fix reference pressure at an interior node
Allow user to fix the reference pressure at an arbitrary location within
the computational domain during simulations.

This capability is important for the pressure Poisson solve during
ABL precursor simulations where the side boundaries are usually set to
periodic and the top boundary is specified with a symmetry BC to allow
the flow to adjust naturally when targeting a specified wind
speed/direction at hub height.