minor cleanup in TpetraLinearSystem for-loops
chatted with Shreyas about the fact that most of the for-loops in
TpetraLinearSystem can't be thread-parallel due to usage of
vector-push-back and other issues.

Refactored most of these loops to just be regular for-loops instead
of being kokkos_parallel_for loops which were hard-wired to use the
serial execution space. Using kokkos_parallel_for was misleading,
giving the impression that they could potentially be run in parallel.