Add ksgs kernel source terms with milestoneRunConsolidated objective
* hook in consolidated kernels. Only valid for Ksgs source terms and
the standard adv/diff/nso flavors. Source terms such as SST and mesh
motion (GCL) are not yest supported.

* milestoneRunConsolidated input file change to activate
lumped mass, adv/diff, and src; diffs in the 1e-16 range; rebless
so that the comparison with SIMD is fair.

* remove forced restart/output in milestone run and push to
threeBlade_rst; forced output based on simulation time
was skewing the timings as a zeroing out of elapsed time occurs.

Notes:

a) at present, tke and z support either consolidated element or not. No mix and match.
This is a design point that we need to figure out as there does seem to be an
added cost to removing nodal contributions for mass and source as compared to
element with lumping. Since we are moving to support vectorized/threaded edge
and nodal loops, perhaps mix-and-matching is in the near future? Performance
will dictate how we proceed. Higher order always needs a full element approach.
As such, we may be required to support nodal and element source terms that do
the same thing - at least from the perspective of a low-order CVFEM scheme.