Fix weird array-bounds warning in Realm.C.
It turns out this is due to a loop-unrolling bug in gcc. It may be
fixed in compiler versions greater than 5, but I'm not positive
about that.

Anyway, the (ugly) fix for us is to hand-unroll the loop over
spatial-dimension. When the compiler unrolls, it makes more
loop iterations than are needed and then some other pass of the
compiler issues the warning about array-bounds even though that
portion of the unrolled loop code is not reachable at runtime.