Add specialized fem-based methods for shape functions extraction from a kernel (#219)
Notes:

a) Not sure if this is the best way. I am following the pattern already
coded and wonder if there is a better way to do all of this. Let me know
in the review:)

b) The original call to scs sized by the scsIp and npe; this probably resuted in a memory
overrun.