Fix bytes_per_thread variable in AssembleElemSolverAlgorithm.
It had been given an overly-conservative multipler of 3, when a
more correct value is 2. During development of the SIMD support,
along with the ScratchViews infrastructure, some memory issues
required the larger allocation to prevent invalid reads/writes.
But those issues are no longer present so the allocation is
now set to the correct value. Verified this by checking the
elemHybridFluids case with valgrind, and also made sure that
all nightly tests pass.