Precursor verification documentation
* New section for precursor verification. I outlined the general use case
in the context of an open-jet simulation study that SNL is running in
partnership with Stanford PSAAP-2.

* Added sub-section on the precursor simulation verification. Towards this
end, I ran a CTV verification problem using a precursor refined CTV
simulation. This data was used for the refinement study. Results indicate
second order accuracy - as expected - when pulling the data from a
external_field_provider realm.

* Added a section for the ABL use case. Demonstrated precursor and subsequent
simulation run with recycling of the precursor data.

* Code change to allow shell-like 2D elements for the CTV use case.

Notes:

a) Verification test is captured in the ABL use case. I probably will not add
the CTV since it is duplicated in nature.